requirements.txt
aiohttp==3.8.5
argparse==1.4.0
asyncio==3.4.3
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0